Cleanup clear and write
