+++ lib/actions/focus.js
@@ -6,7 +6,7 @@ const { highlightElement } = require('../elements/elementHelper');
 const { defaultConfig } = require('../config');
 const { doActionAwaitingNavigation } = require('../doActionAwaitingNavigation');
 
-const focus = async (selector, options, highlight = false) => {
+const focus = async (selector, options = {}, highlight = false) => {
   let elem = selector;
   if (isSelector(selector) || isElement(selector) || isString(selector)) {
     elem = await waitAndGetActionableElement(selector);
