+++ lib/actions/pageActionChecks.js
@@ -8,8 +8,13 @@ const { scrollToElement } = require('./scrollTo');
 const checkVisible = async (elem) => {
   return await elem.isVisible();
 };
+
 const checkNotDisabled = async (elem) => {
   return !(await elem.isDisabled());
+};
+
+const checkWritable = async (elem) => {
+  return await elem.isWritable();
 };
 
 const checkIfElementIsNotCovered = async (e) => {
@@ -62,6 +67,7 @@ const checksMap = {
   visible: { predicate: checkVisible, error: 'is not visible' },
   disabled: { predicate: checkNotDisabled, error: 'is disabled' },
   covered: { predicate: checkIfElementIsNotCovered, error: 'is covered by other element' },
+  writable: { predicate: checkWritable, error: 'is not writable' },
 };
 
 const checkIfActionable = async (elem, checks) => {
