+++ lib/elementSearch.js
@@ -273,6 +273,22 @@ const findElements = async (selector, tag) => {
   return elements;
 };
 
+const findActiveElement = async () => {
+  const getActiveElement = function () {
+    let activeElement = document.activeElement;
+    const parsedShadowRoots = [];
+    while (activeElement.shadowRoot && !parsedShadowRoots.includes(activeElement.shadowRoot)) {
+      parsedShadowRoots.push(activeElement.shadowRoot);
+      activeElement = activeElement.shadowRoot.activeElement;
+    }
+    return activeElement;
+  };
+
+  const activeElementObjectIds = await runtimeHandler.findElements(getActiveElement);
+  const elements = Element.create(activeElementObjectIds, runtimeHandler);
+  return elements[0];
+};
+
 const getIfExists = (findElements, description, customFuncs = {}) => {
   return async (tag, retryInterval, retryTimeout) => {
     retryInterval = determineRetryInterval(retryInterval);
@@ -304,5 +320,6 @@ module.exports = {
   $function,
   findFirstElement,
   findElements,
+  findActiveElement,
   getIfExists,
 };
