+++ lib/elements/element.js
@@ -72,6 +72,47 @@ class Element {
     return result.value;
   }
 
+  async isWritable() {
+    function getDetailsForWrittable() {
+      return {
+        tagName: this.tagName,
+        isContentEditable: this.isContentEditable,
+        disabled: this.disabled,
+        type: this.type,
+        readOnly: this.readOnly,
+      };
+    }
+
+    let editable, disabled;
+    const result = await this.runtimeHandler.runtimeCallFunctionOn(getDetailsForWrittable, null, {
+      objectId: this.objectId,
+      returnByValue: true,
+    });
+    const activeElementDetails = result ? result.result.value : undefined;
+    if (activeElementDetails) {
+      editable =
+        !activeElementDetails.readOnly &&
+        (['INPUT', 'TEXTAREA', 'SELECT'].includes(activeElementDetails.tagName) ||
+          activeElementDetails.isContentEditable);
+      disabled = activeElementDetails.disabled;
+      return !(!editable || disabled);
+    }
+  }
+
+  async isPassword() {
+    function isPassword() {
+      return this.type === 'password';
+    }
+    const { result } = await this.runtimeHandler.runtimeCallFunctionOn(isPassword, null, {
+      objectId: this.objectId,
+      returnByValue: true,
+    });
+    if (result.value === undefined) {
+      return false;
+    }
+    return result.value;
+  }
+
   async isDraggable() {
     function isDraggable() {
       return (this.parentElement && this.parentElement.draggable) || this.draggable;
