+++ lib/handlers/runtimeHandler.js
@@ -161,57 +161,8 @@ const findElements = async (exp, arg) => {
   return objectIds;
 };
 
-const activeElement = async () => {
-  function getActiveElement() {
-    let activeElement = document.activeElement;
-    const parsedShadowRoots = [];
-    while (activeElement.shadowRoot && !parsedShadowRoots.includes(activeElement.shadowRoot)) {
-      parsedShadowRoots.push(activeElement.shadowRoot);
-      activeElement = activeElement.shadowRoot.activeElement;
-    }
-    return activeElement;
-  }
-
-  function getDetails() {
-    return {
-      tagName: this.tagName,
-      isContentEditable: this.isContentEditable,
-      disabled: this.disabled,
-      type: this.type,
-      readOnly: this.readOnly,
-    };
-  }
-
-  let editable, disabled;
-
-  const activeElementObjectIds = await findElements(getActiveElement);
-  for (let activeElementObjectId of activeElementObjectIds) {
-    const result = await runtimeCallFunctionOn(getDetails, null, {
-      objectId: activeElementObjectId,
-      returnByValue: true,
-    });
-    const activeElementDetails = result ? result.result.value : undefined;
-    if (activeElementDetails) {
-      editable =
-        !activeElementDetails.readOnly &&
-        (['INPUT', 'TEXTAREA', 'SELECT'].includes(activeElementDetails.tagName) ||
-          activeElementDetails.isContentEditable);
-      disabled = activeElementDetails.disabled;
-      if (!(!editable || disabled)) {
-        return {
-          notWritable: false,
-          objectId: activeElementObjectId,
-          isPassword: activeElementDetails.type === 'password',
-        };
-      }
-    }
-  }
-  return { notWritable: true };
-};
-
 module.exports = {
   findElements,
-  activeElement,
   runtimeEvaluate,
   runtimeCallFunctionOn,
 };
