+++ lib/taiko.js
@@ -24,7 +24,6 @@ const { description } = require('./actions/pageActionChecks');
 const { match, $$xpath, findElements, findFirstElement } = require('./elementSearch');
 const { handleRelativeSearch, RelativeSearchElement } = require('./proximityElementSearch');
 const Element = require('./elements/element');
-const keyDefinitions = require('../lib/data/USKeyboardLayout');
 const { scrollToElement } = require('./actions/scrollTo');
 const {
   setConfig,
@@ -1039,86 +1038,11 @@ module.exports.focus = async (selector, options = {}) => {
  * @returns {Promise<void>}
  */
 module.exports.write = async (text, into, options = { delay: 0 }) => {
-  if (text == null || text == undefined) {
-    console.warn(`Invalid text value ${text}, setting value to empty ''`);
-    text = '';
-  } else {
-    if (!isString(text)) {
-      text = text.toString();
-    }
-  }
   validate();
-  if (into && !isSelector(into) && !isElement(into) && !isString(into)) {
-    if (!into.delay) {
-      into.delay = options.delay;
-    }
-    options = into;
-    into = undefined;
-  }
+  const { write } = require('./actions/write');
   options = setNavigationOptions(options);
-  let desc, elems, activeElement, elementDesc;
-  if (into) {
-    const selector = isString(into) ? textBox(into) : into;
-    elems = await handleRelativeSearch(await findElements(selector), []);
-    elementDesc = description(selector, true);
-  } else {
-    elems = [{ focused: true }];
-  }
-
-  await waitUntil(
-    async () => {
-      for (let elem of elems) {
-        try {
-          if (!elem.focused) {
-            await focus(elem, options);
-          }
-          activeElement = await runtimeHandler.activeElement();
-          if (activeElement.notWritable) {
-            continue;
-          }
-          return true;
-        } catch (e) {
-          if (e.message.match(/Browser process with pid \d+ exited with/)) {
-            throw e;
-          }
-        }
-      }
-      return false;
-    },
-    defaultConfig.retryInterval,
-    defaultConfig.retryTimeout,
-  ).catch(() => {
-    throw new Error('Element focused is not writable');
-  });
-
-  await doActionAwaitingNavigation(options, async () => {
-    await _write(activeElement, text, options);
-  });
-
-  const textDesc = activeElement.isPassword || options.hideText ? '*****' : text;
-  desc = into
-    ? `Wrote ${textDesc} into the ${elementDesc === '' ? 'focused element' : elementDesc}`
-    : `Wrote ${textDesc} into the focused element.`;
+  const desc = await write(text, into, options);
   descEvent.emit('success', desc);
-};
-
-const _write = async (activeElement, text, options) => {
-  if (defaultConfig.headful) {
-    await highlightElement(new Element(activeElement.objectId, '', runtimeHandler));
-  }
-  for (const char of text) {
-    if (keyDefinitions[char]) {
-      await inputHandler.down(char);
-      await inputHandler.up(char);
-      if (options && options.delay) {
-        await new Promise((resolve) => {
-          setTimeout(resolve, options.delay);
-        });
-      }
-    } else {
-      await inputHandler.sendCharacter(char);
-    }
-  }
 };
 
 /**
@@ -1147,47 +1071,9 @@ module.exports.clear = async (selector, options = {}) => {
     selector = undefined;
   }
   options = setNavigationOptions(options);
-  let activeElement = await runtimeHandler.activeElement();
-  if (activeElement.notWritable || selector) {
-    await waitUntil(
-      async () => {
-        try {
-          if (selector) {
-            await focus(selector, options);
-          }
-        } catch (_) {}
-        return !(await runtimeHandler.activeElement()).notWritable;
-      },
-      100,
-      10000,
-    ).catch(() => {
-      throw new Error('Element cannot be cleared');
-    });
-    activeElement = await runtimeHandler.activeElement();
-  }
-  if (activeElement.notWritable) {
-    throw new Error('Element cannot be cleared');
-  }
-  const desc = !selector ? 'Cleared element on focus' : 'Cleared ' + description(selector, true);
-  await doActionAwaitingNavigation(options, async () => {
-    await _clear(activeElement.objectId);
-    if (defaultConfig.headful) {
-      await highlightElement(new Element(activeElement.objectId, '', runtimeHandler));
-    }
-  });
+  const { clear } = require('./actions/clear');
+  const desc = await clear(selector, options);
   descEvent.emit('success', desc);
-};
-
-const _clear = async (elem) => {
-  await runtimeHandler.runtimeCallFunctionOn(
-    function () {
-      document.execCommand('selectall', false, null);
-    },
-    null,
-    { objectId: elem },
-  );
-  await inputHandler.down('Backspace');
-  await inputHandler.up('Backspace');
 };
 
 /**
